+++ lib/mcapi/crypto/field-level-crypto.js
@@ -258,7 +258,7 @@ function isValidConfig(config) {
     throw Error("Config not valid: dataEncoding should be 'hex' or 'base64'");
   }
   validateFingerprint(config, contains);
-  validateRootMapping(config);
+  utils.validateRootMapping(config);
 }
 
 function validateFingerprint(config, contains) {
@@ -278,25 +278,6 @@ function validateFingerprint(config, contains) {
       "Config not valid: propertiesFingerprint should be: 'certificate' or 'publicKey'"
     );
   }
-}
-
-function validateRootMapping(config) {
-  function multipleRoots(elem) {
-    return (
-      elem.length !== 1 &&
-      elem.some((item) => {
-        return item.obj === "$" || item.element === "$";
-      })
-    );
-  }
-
-  config.paths.forEach((path) => {
-    if (multipleRoots(path.toEncrypt) || multipleRoots(path.toDecrypt)) {
-      throw Error(
-        "Config not valid: found multiple configurations encrypt/decrypt with root mapping"
-      );
-    }
-  });
 }
 
 module.exports = FieldLevelCrypto;
