+++ lib/mcapi/crypto/jwe-crypto.js
@@ -246,7 +246,7 @@ function isValidConfig(config) {
     throw Error("Config not valid: paths should be not empty.");
   }
   validateFingerprint(config, contains);
-  validateRootMapping(config);
+  utils.validateRootMapping(config);
 }
 
 /**
@@ -264,28 +264,6 @@ function validateFingerprint(config, contains) {
       "Config not valid: propertiesFingerprint should be: 'certificate' or 'publicKey'"
     );
   }
-}
-
-/**
- * @private
- */
-function validateRootMapping(config) {
-  function multipleRoots(elem) {
-    return (
-      elem.length !== 1 &&
-      elem.some((item) => {
-        return item.obj === "$" || item.element === "$";
-      })
-    );
-  }
-
-  config.paths.forEach((path) => {
-    if (multipleRoots(path.toEncrypt) || multipleRoots(path.toDecrypt)) {
-      throw Error(
-        "Config not valid: found multiple configurations encrypt/decrypt with root mapping"
-      );
-    }
-  });
 }
 
 /**
