+++ lib/mcapi/encryption/field-level-encryption.js
@@ -25,43 +25,6 @@ function FieldLevelEncryption(config) {
 }
 
 /**
- * @private
- */
-function hasConfig(config, endpoint) {
-  if (config && endpoint) {
-    endpoint = endpoint.split("?").shift();
-    const conf = config.paths.find((elem) => {
-      const regex = new RegExp(elem.path, "g");
-      return endpoint.match(regex);
-    });
-    return conf ? conf : null;
-  }
-  return null;
-}
-
-/**
- * @private
- */
-function elemFromPath(path, obj) {
-  try {
-    let parent = null;
-    const paths = path.split(".");
-    if (path && paths.length > 0) {
-      paths.forEach((e) => {
-        parent = obj;
-        obj = isJsonRoot(e) ? obj : obj[e];
-      });
-    }
-    return {
-      node: obj,
-      parent: parent,
-    };
-  } catch (e) {
-    return null;
-  }
-}
-
-/**
  * Set encryption header parameters
  *
  * @param header HTTP header
@@ -86,7 +49,7 @@ function setHeader(header, params) {
  */
 function encrypt(endpoint, header, body) {
   let bodyMap = body;
-  const fleConfig = hasConfig(this.config, endpoint);
+  const fleConfig = utils.hasConfig(this.config, endpoint);
   if (fleConfig) {
     if (!this.isWithHeader) {
       bodyMap = fleConfig.toEncrypt.map((v) => {
@@ -102,7 +65,9 @@ function encrypt(endpoint, header, body) {
   }
   return {
     header: header,
-    body: fleConfig ? computeBody(fleConfig.toEncrypt, body, bodyMap) : body,
+    body: fleConfig
+      ? utils.computeBody(fleConfig.toEncrypt, body, bodyMap)
+      : body,
   };
 }
 
@@ -116,7 +81,7 @@ function encrypt(endpoint, header, body) {
 function decrypt(response) {
   const body = response.body;
   let bodyMap = response.body;
-  const fleConfig = hasConfig(this.config, response.request.url);
+  const fleConfig = utils.hasConfig(this.config, response.request.url);
   if (fleConfig) {
     bodyMap = fleConfig.toDecrypt.map((v) => {
       if (!this.isWithHeader) {
@@ -126,7 +91,9 @@ function decrypt(response) {
       }
     });
   }
-  return fleConfig ? computeBody(fleConfig.toDecrypt, body, bodyMap) : body;
+  return fleConfig
+    ? utils.computeBody(fleConfig.toDecrypt, body, bodyMap)
+    : body;
 }
 
 /**
@@ -137,13 +104,13 @@ function decrypt(response) {
  * @param body Body to encrypt
  */
 function encryptBody(path, body) {
-  const elem = elemFromPath(path.element, body);
+  const elem = utils.elemFromPath(path.element, body);
   if (elem && elem.node) {
     const encryptedData = this.crypto.encryptData({ data: elem.node });
     body = utils.mutateObjectProperty(path.obj, encryptedData, body);
     // delete encrypted field if not overridden
     if (
-      !isJsonRoot(path.obj) &&
+      !utils.isJsonRoot(path.obj) &&
       path.element !== path.obj + "." + this.config.encryptedValueFieldName
     ) {
       utils.deleteNode(path.element, body);
@@ -162,13 +129,13 @@ function encryptBody(path, body) {
  * @returns {Object} Encrypted body
  */
 function encryptWithHeader(encParams, path, body) {
-  const elem = elemFromPath(path.element, body).node;
+  const elem = utils.elemFromPath(path.element, body).node;
   const encrypted = this.crypto.encryptData({ data: elem }, encParams);
   const data = {
     [this.config.encryptedValueFieldName]:
       encrypted[this.config.encryptedValueFieldName],
   };
-  return isJsonRoot(path.obj) ? data : { [path.obj]: data };
+  return utils.isJsonRoot(path.obj) ? data : { [path.obj]: data };
 }
 
 /**
@@ -180,7 +147,7 @@ function encryptWithHeader(encParams, path, body) {
  * @returns {Object} Decrypted body
  */
 function decryptBody(path, body) {
-  const elem = elemFromPath(path.element, body);
+  const elem = utils.elemFromPath(path.element, body);
   if (elem && elem.node) {
     const decryptedObj = this.crypto.decryptData(
       elem.node[this.config.encryptedValueFieldName], // encrypted data
@@ -208,8 +175,8 @@ function decryptBody(path, body) {
  * @param response Response with header to update
  */
 function decryptWithHeader(path, body, response) {
-  const elemEncryptedNode = elemFromPath(path.obj, body);
-  const node = isJsonRoot(path.element)
+  const elemEncryptedNode = utils.elemFromPath(path.obj, body);
+  const node = utils.isJsonRoot(path.element)
     ? elemEncryptedNode.node
     : elemEncryptedNode.node[path.element];
   if (node) {
@@ -224,20 +191,10 @@ function decryptWithHeader(path, body, response) {
       response.header[this.config.oaepHashingAlgorithmHeaderName],
       response.header[this.config.encryptedKeyHeaderName]
     );
-    return isJsonRoot(path.obj) ? decrypted : Object.assign(body, decrypted);
+    return utils.isJsonRoot(path.obj)
+      ? decrypted
+      : Object.assign(body, decrypted);
   }
-}
-
-function isJsonRoot(elem) {
-  return elem === "$";
-}
-
-function computeBody(configParam, body, bodyMap) {
-  return hasEncryptionParam(configParam, bodyMap) ? bodyMap.pop() : body;
-}
-
-function hasEncryptionParam(encParams, bodyMap) {
-  return encParams && encParams.length === 1 && bodyMap && bodyMap[0];
 }
 
 module.exports = FieldLevelEncryption;
