+++ lib/mcapi/encryption/jwe-encryption.js
@@ -16,43 +16,6 @@ function JweEncryption(config) {
 }
 
 /**
- * @private
- */
-function hasConfig(config, endpoint) {
-  if (config && endpoint) {
-    endpoint = endpoint.split("?").shift();
-    const conf = config.paths.find((elem) => {
-      const regex = new RegExp(elem.path, "g");
-      return endpoint.match(regex);
-    });
-    return conf ? conf : null;
-  }
-  return null;
-}
-
-/**
- * @private
- */
-function elemFromPath(path, obj) {
-  try {
-    let parent = null;
-    const paths = path.split(".");
-    if (path && paths.length > 0) {
-      paths.forEach((e) => {
-        parent = obj;
-        obj = isJsonRoot(e) ? obj : obj[e];
-      });
-    }
-    return {
-      node: obj,
-      parent: parent,
-    };
-  } catch (e) {
-    return null;
-  }
-}
-
-/**
  * Encrypt parts of a HTTP request using the given config
  *
  * @param endpoint HTTP URL for the current call
@@ -62,7 +25,7 @@ function elemFromPath(path, obj) {
  */
 function encrypt(endpoint, header, body) {
   let bodyMap = body;
-  const fleConfig = hasConfig(this.config, endpoint);
+  const fleConfig = utils.hasConfig(this.config, endpoint);
   if (fleConfig) {
     bodyMap = fleConfig.toEncrypt.map((v) => {
       return encryptBody.call(this, v, body);
@@ -70,7 +33,9 @@ function encrypt(endpoint, header, body) {
   }
   return {
     header: header,
-    body: fleConfig ? computeBody(fleConfig.toEncrypt, body, bodyMap) : body,
+    body: fleConfig
+      ? utils.computeBody(fleConfig.toEncrypt, body, bodyMap)
+      : body,
   };
 }
 
@@ -84,13 +49,15 @@ function encrypt(endpoint, header, body) {
 function decrypt(response) {
   const body = response.body;
   let bodyMap = response.body;
-  const fleConfig = hasConfig(this.config, response.request.url);
+  const fleConfig = utils.hasConfig(this.config, response.request.url);
   if (fleConfig) {
     bodyMap = fleConfig.toDecrypt.map((v) => {
       return decryptBody.call(this, v, body);
     });
   }
-  return fleConfig ? computeBody(fleConfig.toDecrypt, body, bodyMap) : body;
+  return fleConfig
+    ? utils.computeBody(fleConfig.toDecrypt, body, bodyMap)
+    : body;
 }
 
 /**
@@ -101,13 +68,13 @@ function decrypt(response) {
  * @param body Body to encrypt
  */
 function encryptBody(path, body) {
-  const elem = elemFromPath(path.element, body);
+  const elem = utils.elemFromPath(path.element, body);
   if (elem && elem.node) {
     const encryptedData = this.crypto.encryptData({ data: elem.node });
     body = utils.mutateObjectProperty(path.obj, encryptedData, body);
     // delete encrypted field if not overridden
     if (
-      !isJsonRoot(path.obj) &&
+      !utils.isJsonRoot(path.obj) &&
       path.element !== path.obj + "." + this.config.encryptedValueFieldName
     ) {
       utils.deleteNode(path.element, body);
@@ -125,7 +92,7 @@ function encryptBody(path, body) {
  * @returns {Object} Decrypted body
  */
 function decryptBody(path, body) {
-  const elem = elemFromPath(path.element, body);
+  const elem = utils.elemFromPath(path.element, body);
   if (elem && elem.node) {
     const decryptedObj = this.crypto.decryptData(
       elem.node[this.config.encryptedValueFieldName]
@@ -139,18 +106,6 @@ function decryptBody(path, body) {
     );
   }
   return body;
-}
-
-function isJsonRoot(elem) {
-  return elem === "$";
-}
-
-function computeBody(configParam, body, bodyMap) {
-  return hasEncryptionParam(configParam, bodyMap) ? bodyMap.pop() : body;
-}
-
-function hasEncryptionParam(encParams, bodyMap) {
-  return encParams && encParams.length === 1 && bodyMap && bodyMap[0];
 }
 
 module.exports = JweEncryption;
