+++ lib/mcapi/utils/utils.js
@@ -312,3 +312,68 @@ module.exports.nodeVersionSupportsJWE = function () {
   const nodeMajorVersion = parseInt(process.version.substring(1, 3));
   return nodeMajorVersion > 12;
 };
+
+module.exports.hasConfig = function (config, endpoint) {
+  if (config && endpoint) {
+    endpoint = endpoint.split("?").shift();
+    const conf = config.paths.find((elem) => {
+      const regex = new RegExp(elem.path, "g");
+      return endpoint.match(regex);
+    });
+    return conf ? conf : null;
+  }
+  return null;
+};
+
+module.exports.elemFromPath = function (path, obj) {
+  try {
+    let parent = null;
+    const paths = path.split(".");
+    if (path && paths.length > 0) {
+      paths.forEach((e) => {
+        parent = obj;
+        obj = isJsonRoot(e) ? obj : obj[e];
+      });
+    }
+    return {
+      node: obj,
+      parent: parent,
+    };
+  } catch (e) {
+    return null;
+  }
+};
+
+module.exports.isJsonRoot = function (elem) {
+  return isJsonRoot(elem);
+};
+function isJsonRoot(elem) {
+  return elem === "$";
+}
+
+module.exports.computeBody = function (configParam, body, bodyMap) {
+  return hasEncryptionParam(configParam, bodyMap) ? bodyMap.pop() : body;
+};
+
+function hasEncryptionParam(encParams, bodyMap) {
+  return encParams && encParams.length === 1 && bodyMap && bodyMap[0];
+}
+
+module.exports.validateRootMapping = function (config) {
+  function multipleRoots(elem) {
+    return (
+      elem.length !== 1 &&
+      elem.some((item) => {
+        return item.obj === "$" || item.element === "$";
+      })
+    );
+  }
+
+  config.paths.forEach((path) => {
+    if (multipleRoots(path.toEncrypt) || multipleRoots(path.toDecrypt)) {
+      throw Error(
+        "Config not valid: found multiple configurations encrypt/decrypt with root mapping"
+      );
+    }
+  });
+};
